Refactor project structure: Move StationData to types directory
- Relocate StationData interfaces to a new /src/types directory
- Update import statements across components and services
- Remove unused station name lookup function from stationService
- Add comprehensive JSDoc comments to improve type documentation
