+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import { fetchStationData } from "@/services/stationService";
-import { StationData } from "@/services/StationData";
+import { StationData } from "@/types/StationData";
 import { useState } from "react";
 import StationSearch from "@/components/StationSearch";
 export default function Home() {
